pyproject.toml
[project]
package-mode = false
name = "parser-bet"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "playwright (>=1.53.0,<2.0.0)",
    "nicegui (>=2.21.1,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "pandas (>=2.3.1,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "playwright-stealth (>=2.0.0,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "lxml (>=6.0.0,<7.0.0)",
    "html5lib (>=1.1,<2.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
flake8 = "^7.3.0"
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
flake8-pyproject = "^1.2.3"
pudb = "^2025.1"
poetry-plugin-dotenv = "^3.0.1"


[tool.flake8]
max_line_length = 120

